'Solution'

+++ car/serializers.py
@@ -6,7 +6,7 @@ from car.models import Car
 class CarSerializer(serializers.Serializer):
     manufacturer = serializers.CharField(max_length=64)
     model = serializers.CharField(max_length=64)
-    horse_powers = serializers.PositiveSmallIntegerField(
+    horse_powers = serializers.PositiveIntegerField(
         min_value=1,
         max_value=1914
     )
